09-17-23 bug fixes

+++ genal/__init__.py
@@ -18,8 +18,8 @@ config_path = os.path.join(config_dir, "config.json")
 if not os.path.exists(config_dir):
     os.makedirs(config_dir)
     
-#if not os.path.exists(config_path):
-default_config = default_config()
-with open(config_path, "w") as f:
-    json.dump(default_config, f)
-print (f"Configuration file for genal placed at '{config_path}'")
+if not os.path.exists(config_path):
+    default_config = default_config()
+    with open(config_path, "w") as f:
+        json.dump(default_config, f)
+    print (f"Configuration file for genal placed at '{config_path}'")
